==> 06_Cell_cell_communication_analysis.Rmd <==
## Load required packages
```{r load_pkgs, message=FALSE, warning=FALSE}
library(tidyverse)
library(magrittr)
library(liana)
library(Seurat)
```

## CCC Resources
```{r liana_resource, warning = FALSE}
# Resource currently included in OmniPathR (and hence `liana`) include:
show_resources()
```
  
## CCC Methods
```{r liana_method, warning = FALSE}
# Resource currently included in OmniPathR (and hence `liana`) include:
show_methods()
```

```{r load_data}
wd <- "/raida/ayaka.mori/integrative_dp/rna_preprocessing/preprocessing_output_simple"
obj <- readRDS(paste0(wd, "/scalp_fib.rds"))
```

```{r}
obj_aa <- subset(obj, subset = diseaseStatus == "AA")
obj_cpb <- subset(obj, subset = diseaseStatus == "C_PB")
```

#aa
```{r}
DimPlot(obj_aa, reduction = "umap", label = TRUE) + NoLegend()
```

```{r}
obj_aa <- SetIdent(obj_aa, value = "clust_chat_new")
DimPlot(obj_aa, reduction = "umap", label = TRUE) + NoLegend()
```

```{r}
# Run liana
liana_test_aa <- liana_wrap(obj_aa)
# Liana returns a list of results, each element of which corresponds to a method
liana_test_aa %>% dplyr::glimpse()
```

```{r}
# We can aggregate these results into a tibble with consensus ranks
liana_test_aa <- liana_test_aa %>%
  liana_aggregate()

dplyr::glimpse(liana_test_aa)
```

```{r}
cellchat_test_aa <- liana_wrap(obj_aa, method = "call_cellchat")
# Liana returns a list of results, each element of which corresponds to a method
cellchat_test_aa %>% dplyr::glimpse()
```

```{r}
colnames(cellchat_test_aa) <- c("source", "target", "ligand.complex", "receptor.complex", "cellchat.prob", "cellchat.pval")
```

```{r}
liana_aa_merge <- merge(liana_test_aa, cellchat_test_aa, by = c("source", "target", "ligand.complex", "receptor.complex"))
```

```{r}
write.csv(liana_aa_merge, file = "liana_aa_merge.csv", row.names = FALSE)
```

#cpb
```{r}
obj_cpb <- SetIdent(obj_cpb, value = "clust_chat_new")
DimPlot(obj_cpb, reduction = "umap", label = TRUE) + NoLegend()
```
```{r}
# Run liana
liana_test_cpb <- liana_wrap(obj_cpb)
# Liana returns a list of results, each element of which corresponds to a method
liana_test_cpb %>% dplyr::glimpse()
```

```{r}
# We can aggregate these results into a tibble with consensus ranks
liana_test_cpb <- liana_test_cpb %>%
  liana_aggregate()

dplyr::glimpse(liana_test_cpb)
```

```{r}
cellchat_test_cpb <- liana_wrap(obj_cpb, method = "call_cellchat")
# Liana returns a list of results, each element of which corresponds to a method
cellchat_test_cpb %>% dplyr::glimpse()
```

```{r}
colnames(cellchat_test_cpb) <- c("source", "target", "ligand.complex", "receptor.complex", "cellchat.prob", "cellchat.pval")
```

```{r}
liana_cpb_merge <- merge(liana_test_cpb, cellchat_test_cpb, by = c("source", "target", "ligand.complex", "receptor.complex"))
```

```{r}
write.csv(liana_cpb_merge, file = "liana_cpb_merge.csv", row.names = FALSE)
```
