==> 07_VennDiagram_LR.Rmd <==
```{r}
#Load package
library(VennDiagram)
```

```{r}
#load LR-data
LR_growth <- read.csv("hair_growth_LR.csv", header = TRUE)
LR_loss <- read.csv("hair_loss_LR.csv", header = TRUE)
```

```{r}
#load LR-data
LR_CellPhoneDB_growth <- read.csv("CellPhoneDB_cpb.csv", header = TRUE)
LR_CellChat_growth <- read.csv("CellChat_cpb.csv", header = TRUE)

LR_CellPhoneDB_loss <- read.csv("CellPhoneDB_aa.csv", header = TRUE)
LR_CellChat_loss <- read.csv("CellChat_aa.csv", header = TRUE)
```

```{r}
# Prepare a palette of 3 colors with R colorbrewer:
library(RColorBrewer)
myCol <- brewer.pal(3, "Pastel2")
```

```{r}
#hair growth-related LR
DEGs_growth <- LR_growth$hair.growth.related.genes
CellPhone_growth <- LR_CellPhoneDB_growth$CellPhoneDB
CellChat_growth <- LR_CellChat_growth$CellChat
data_hair_growth_LR <- list("DEGs" = DEGs_growth, "CellPhoneDB" = CellPhone_growth, "CellChat" = CellChat_growth)
venn_LR_growth <-venn.diagram(data_hair_growth_LR, 
             filename = NULL,
             width = 2000, height = 2000,
             resolution = 300,
             # Circles
             lwd = 2,
             lty = 'blank',
             fill = myCol,
             # Numbers
             cex = 1,
             fontface = "plain",
             fontfamily = "sans",
             # Set names
             cat.cex = 1,
             cat.fontface = "plain",
             cat.default.pos = "outer",
             cat.fontfamily = "sans")
library(grDevices)
grid.draw(venn_LR_growth)
```

```{r}
#hair loss-related LR
DEGs_loss <- LR_loss$hair.loss.related.genes
CellPhone_loss <- LR_CellPhoneDB_loss$CellPhoneDB
CellChat_loss <- LR_CellChat_growth$CellChat
data_hair_loss_LR <- list("DEGs" = DEGs_loss, "CellPhoneDB" = CellPhone_loss, "CellChat" = CellChat_loss)
venn_LR_loss <-venn.diagram(data_hair_loss_LR, 
             filename = NULL,
             width = 2000, height = 2000,
             resolution = 300,
             # Circles
             lwd = 2,
             lty = 'blank',
             fill = myCol,
             # Numbers
             cex = 1,
             fontface = "plain",
             fontfamily = "sans",
             # Set names
             cat.cex = 1,
             cat.fontface = "plain",
             cat.default.pos = "outer",
             cat.fontfamily = "sans")
grid.draw(venn_LR_loss)
```
